IPiece
Added and wired the IPiece interface.

Moved the functionality of changing the position of the sides to the
sides instead of the piece.

Added coordinates to the pieces that will be used instead of where they
are in an array.

Changed the 3-D array to a list of IPiece.

Remove a function that was no longer needed.

Added a lastMove to allow the undoing of moves.

Added a way to get every piece in a certain layer on a certain axis.
This allows access to all pieces that need to be moved.