Removed reorientation
Instead of orienting after every move, the corner at 000 will never
move.

Fixed a couple of bug for changing the direction when trying to move 000
or when undoing a move.

Trying to move a layer that doesn't exist throws an error.

Fixed logic for moving pieces to their next coordinates.

Started using NCrunch for continuous testing.