Solver & ICube
Created class for the solver, but nothing is in it.

Created an ICube interface, but have yet to figure out what will go in
it.

Added a path property to the cube in order to record the steps taken.

Added a Step class that is the representation of each step of that cube.